Flatten react widgets example layout
