+++ examples/widgets/react-widgets/index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="./src/main.tsx"></script>
+    <script type="module" src="./index.tsx"></script>
   </body>
 </html>
