+++ examples/widgets/react-widgets/index.tsx
@@ -1,7 +1,11 @@
+// deck.gl-community
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
 import React from 'react';
 import {createRoot} from 'react-dom/client';
 
-import {App} from './example';
+import {App} from './app';
 
 const container = document.getElementById('app');
 if (!container) {
